Separate protocols into files. Create new package
